requirements.txt
flask==3.0.0
anthropic==0.34.2
python-dotenv==1.0.0
